Move nth value to its own file
